Add files via upload
Code for looking through AmStories hits (based on kw proximity) with GPT-cleaned text to find kw hits and unique instances of the string 'explor'.  This produces a df without all of the metadata from the original AmStories entry based on the notion that it can be looked up again using an older intermediate work csv.  This should produces a df that should be ready for content evaluation.